test(tui): split scrollback coverage into its own module
The rewrite pushed test_tui_event_rendering.py to 806 lines, six over the
800-line ceiling. On a pull request that ceiling only measures files the branch
adds, so this would have gone green here and broken the push-side run on main.

The seam is the one the rewrite created: event dispatch and per-agent routing
stay put, while the printing contract — which blocks reach the terminal, when,
and exactly once — moves to test_tui_scrollback.py.
